Separate Localization feature into a separate entity and component
