+++ public/locales/en/common.json
@@ -1,0 +1,5 @@
+{
+  "language": "Language",
+  "en": "English 🇬🇧",
+  "ru": "Russian 🇷🇺"
+}
