+++ public/locales/en/settings.json
@@ -7,8 +7,5 @@
   "huge": "Huge",
   "controls": "Controls",
   "on": "On",
-  "off": "Off",
-  "language": "Language",
-  "en": "English 🇬🇧",
-  "ru": "Russian 🇷🇺"
+  "off": "Off"
 }
