+++ public/locales/ru/common.json
@@ -1,0 +1,5 @@
+{
+  "language": "Язык",
+  "en": "Английский 🇬🇧",
+  "ru": "Русский 🇷🇺"
+}
