+++ public/locales/ru/settings.json
@@ -7,8 +7,5 @@
   "huge": "Огромный",
   "controls": "Управление",
   "on": "Вкл",
-  "off": "Выкл",
-  "language": "Язык",
-  "en": "Английский 🇬🇧",
-  "ru": "Русский 🇷🇺"
+  "off": "Выкл"
 }
