+++ src/features/localization/index.ts
@@ -1,0 +1,5 @@
+import LanguageSwitcher from "./components/LanguageSwitcher";
+import {useLanguage} from "./hooks/useLanguage";
+import initI18n from "./i18n";
+
+export {LanguageSwitcher, useLanguage, initI18n};
